Refactor API models: reorganize WeatherForecast and WelcomeResponse classes, and update namespace usage
