+++ src/API/Controllers/WeatherForecastController.cs
@@ -1,3 +1,4 @@
+using API.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers;
