+++ src/API/Models/WeatherForecast.cs
@@ -1,4 +1,4 @@
-namespace API;
+namespace API.Models;
 
 public class WeatherForecast
 {
