+++ src/API/Models/WelcomeResponse.cs
@@ -1,0 +1,9 @@
+namespace API.Models;
+
+/// <summary>
+/// Response model for the welcome endpoint
+/// </summary>
+public sealed record WelcomeResponse(
+    string Message,
+    DateTime Timestamp,
+    Guid RequestId);
